# openapi.yaml
openapi: 3.1.0
info:
  title: 企业短信管理平台 API
  version: 1.6.0
  description: |
    统一短信发送与管理平台对外接口。
    - /api/v1/messages/** 供内部系统调用，X-Api-Key 认证（双 Key 轮换宽限期内新旧均有效）
    - /api/v1/web/** 供 Vue3 前端调用，显式选择本地或 AD 认证源；15 分钟 access JWT 每次按数据库权威安全上下文校验，refresh JWT 七天内单次轮换
    - 所有时间均为 ISO8601 +08:00
    - v1.1：消息分类(category)、biz_id 幂等、失败重发、结果回调、双 Key 轮换、登出
    - v1.2：计费条(est_segments/quota_cost)、模板变量渲染(template_params)、号码级频控(removed_freq_limit)、
      测试发送(is_test)、营销同意(consent_confirmed)、批次列表/号码搜索/队列恢复/用户管理/回调密钥轮换
    - v1.3-v1.5：厂商{sN}模板转换、OTP打码、unmatched、异常检测、号码时间线、任务健康
    - v1.6：AUTH_MOCK、完整CRUD/计费预估/改期/授权解密、密文导出、owner/app角色与确定性Mock验收
    - v1.6.13：用户角色变更与强制下线使用数据库会话版本即时失效既有 JWT
    - v1.6.14：本地/AD Provider 显式路由、稳定主体 JWT 与首次强制改密
    - v1.6.24：统一 security_version 覆盖账号/身份/Provider/部门/角色，短 access JWT 与单次 refresh 轮换
    - v1.6.12：运行策略事务最终校验、动态过期时间与启动调度心跳一致
    - v1.6.11：报告回调引用新增无 PII 事件身份，raw 重放不重复投递同一报告
    - v1.6.28：callback/export 使用可续租 UUID fencing；回调暴露稳定 event_id 与租约状态
    - v1.6.33：拆分无依赖 livez 与严格依赖 readyz；健康响应不暴露内部组件或失败原因
    - v1.6.34：生产 callback 强制 HTTPS；签名密钥/协议版本按事件固化；敏感密文与导出帧绑定业务上下文
servers:
  - url: https://sms.example.com
...
paths:
...
  /api/v1/web/admin/queue/status:
    get:
      tags: [admin]
      summary: 发送队列暂停原因与余额状态
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: 队列状态
          content: { application/json: { schema: { $ref: '#/components/schemas/QueueStatusModel' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/web/admin/queue/resume:
    post:
      tags: [admin]
      summary: 余额恢复后一键恢复发送队列（v1.2，balance_blocked 批次重新入队）
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: force, in: query, required: false, schema: { type: boolean, default: false }, description: 显式break-glass恢复 }
      responses:
        '200':
          description: 已恢复
          content:
            application/json:
              schema: { $ref: '#/components/schemas/QueueResumeModel' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /api/v1/web/admin/jobs:
    get:
      tags: [admin]
      summary: 后台任务健康（v1.5，各任务最近运行/耗时/24h成功率）
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: 任务列表
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/JobModel' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/web/admin/jobs/{job_name}/trigger:
    post:
      tags: [admin]
      summary: 手动触发一次任务（v1.5，记审计）
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: job_name, in: path, required: true, schema: { type: string } }
      responses:
        '202': { description: 已投递执行 }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/web/admin/alerts:
    get:
      tags: [admin]
      summary: 告警记录列表（v1.4，ops页支撑）
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: alert_type, in: query, required: false, schema: { anyOf: [{ type: string }, { type: 'null' }] } }
        - { name: level, in: query, required: false, schema: { anyOf: [{ type: string, enum: [info, warn, crit] }, { type: 'null' }] } }
        - { name: start, in: query, required: false, schema: { anyOf: [{ type: string, format: date-time }, { type: 'null' }] } }
        - { name: end, in: query, required: false, schema: { anyOf: [{ type: string, format: date-time }, { type: 'null' }] } }
        - { name: page, in: query, required: false, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: page_size, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/AlertPageModel' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/web/admin/outbox:
    get:
      tags: [admin]
      summary: Transactional Outbox 积压、失败与最老事件年龄
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: 无 PII Outbox 运行指标
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OutboxStatsModel' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/web/admin/outbox/{event_id}/retry:
    post:
      tags: [admin]
      summary: 由稳定管理员主体审计后重推 dead Outbox 事件
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: event_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '204': { description: 已恢复为 pending，等待 dispatcher }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/web/admin/raw-logs:
    get:
      tags: [admin]
      summary: 厂商原始报文密文元数据列表（不返回payload_enc或手机号）
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: source, in: query, required: false, schema: { anyOf: [{ type: string, enum: [report, reply] }, { type: 'null' }] } }
        - { name: processed, in: query, required: false, schema: { anyOf: [{ type: boolean }, { type: 'null' }] } }
        - { name: page, in: query, required: false, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: page_size, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/RawLogPageModel' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/web/admin/chunks/uncertain:
    get:
      tags: [admin]
      summary: uncertain 分片列表（v1.4，待人工核查项）
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: page, in: query, required: false, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: page_size, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        '200': { description: 含批次号/customId/停留时长, content: { application/json: { schema: { $ref: '#/components/schemas/UncertainPageModel' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/web/admin/unmatched-reports:
    get:
      tags: [admin]
      summary: 无主报告查询（v1.4，并行迁移期对账；支持导出）
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: phone, in: query, required: false, schema: { anyOf: [{ type: string, pattern: '^1\d{10}$' }, { type: 'null' }] }, description: 精确匹配走HMAC }
        - { name: start, in: query, required: false, schema: { anyOf: [{ type: string, format: date-time }, { type: 'null' }] } }
        - { name: end, in: query, required: false, schema: { anyOf: [{ type: string, format: date-time }, { type: 'null' }] } }
        - { name: page, in: query, required: false, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: page_size, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        '200': { description: 列表(号码mask), content: { application/json: { schema: { $ref: '#/components/schemas/UnmatchedPageModel' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/web/admin/unmatched-reports/export:
    post:
      tags: [admin]
      summary: 创建无主报告对账导出任务（密文文件，下载时流式解密）
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UnmatchedExportModel' }
      responses:
        '202':
          description: 已创建
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExportTask' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/web/admin/raw-logs/{id}/replay:
    post:
      tags: [admin]
      summary: 重放厂商原始报文（v1.1，解析失败兜底）
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: 已重放解析
          content: { application/json: { schema: { $ref: '#/components/schemas/ReplayResultModel' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    MetricsBearerAuth:
      type: http
      scheme: bearer
      description: 仅供 Prometheus 抓取使用的独立、可轮换 secret。
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: 参数错误
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Unauthorized:
      description: 认证失败
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Forbidden:
      description: 权限不足或审批回避
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Conflict:
      description: 状态冲突
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound:
      description: 资源不存在
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  schemas:
...
    AlertModel:
      type: object
      required: [id, alert_type, level, title, detail, channels, created_at]
      properties:
        id: { type: integer }
        alert_type: { type: string }
        level: { type: string, enum: [info, warn, crit] }
        title: { type: string }
        detail: { anyOf: [{ type: object, additionalProperties: true }, { type: 'null' }] }
        channels: { type: string }
        created_at: { type: string, format: date-time }
    AlertPageModel:
      type: object
      required: [total, page, page_size, items]
      properties:
        total: { type: integer, minimum: 0 }
        page: { type: integer, minimum: 1 }
        page_size: { type: integer, minimum: 1, maximum: 100 }
        items: { type: array, items: { $ref: '#/components/schemas/AlertModel' } }
    RawLogModel:
      type: object
      required: [id, source, item_count, custom_id_count, processed, error, fetched_at]
      properties:
        id: { type: integer }
        source: { type: string, enum: [report, reply] }
        item_count: { type: integer }
        custom_id_count: { type: integer }
        processed: { type: boolean }
        error: { anyOf: [{ type: string }, { type: 'null' }] }
        fetched_at: { type: string, format: date-time }
    RawLogPageModel:
      type: object
      required: [total, page, page_size, items]
      properties:
        total: { type: integer, minimum: 0 }
        page: { type: integer, minimum: 1 }
        page_size: { type: integer, minimum: 1, maximum: 100 }
        items: { type: array, items: { $ref: '#/components/schemas/RawLogModel' } }
    UncertainModel:
      type: object
      required: [chunk_id, batch_no, custom_id, phone_count, vendor_code, uncertain_since, age_seconds]
      properties:
        chunk_id: { type: integer }
        batch_no: { type: string }
        custom_id: { type: string }
        phone_count: { type: integer }
        vendor_code: { anyOf: [{ type: integer }, { type: 'null' }] }
        uncertain_since: { type: string, format: date-time }
        age_seconds: { type: integer }
    UncertainPageModel:
      type: object
      required: [total, page, page_size, items]
      properties:
        total: { type: integer, minimum: 0 }
        page: { type: integer, minimum: 1 }
        page_size: { type: integer, minimum: 1, maximum: 100 }
        items: { type: array, items: { $ref: '#/components/schemas/UncertainModel' } }
    UnmatchedModel:
      type: object
      required: [id, vendor_task_id, custom_id, phone_mask, report_status, report_desc, report_time, created_at]
      properties:
        id: { type: integer }
        vendor_task_id: { anyOf: [{ type: string }, { type: 'null' }] }
        custom_id: { anyOf: [{ type: string }, { type: 'null' }] }
        phone_mask: { type: string }
        report_status: { anyOf: [{ type: integer }, { type: 'null' }] }
        report_desc: { anyOf: [{ type: string }, { type: 'null' }] }
        report_time: { anyOf: [{ type: string, format: date-time }, { type: 'null' }] }
        created_at: { type: string, format: date-time }
    UnmatchedPageModel:
      type: object
      required: [total, page, page_size, items]
      properties:
        total: { type: integer, minimum: 0 }
        page: { type: integer, minimum: 1 }
        page_size: { type: integer, minimum: 1, maximum: 100 }
        items: { type: array, items: { $ref: '#/components/schemas/UnmatchedModel' } }
    UnmatchedExportModel:
      type: object
      properties:
        phone: { anyOf: [{ type: string, pattern: '^1\d{10}$' }, { type: 'null' }] }
        start: { anyOf: [{ type: string, format: date-time }, { type: 'null' }] }
        end: { anyOf: [{ type: string, format: date-time }, { type: 'null' }] }
        decrypted: { type: boolean, default: false }
    JobModel:
      type: object
      required: [job_name, last_run_at, last_status, last_duration_ms, last_items, success_rate_24h, stalled]
      properties:
        job_name: { type: string }
        last_run_at: { anyOf: [{ type: string, format: date-time }, { type: 'null' }] }
        last_status: { anyOf: [{ type: string, enum: [running, success, failed] }, { type: 'null' }] }
        last_duration_ms: { anyOf: [{ type: integer }, { type: 'null' }] }
        last_items: { type: integer }
        success_rate_24h: { type: number }
        stalled: { type: boolean }
    OutboxStatsModel:
      type: object
      required: [pending, published, processing, dead, failed_attempts, oldest_age_seconds]
      properties:
        pending: { type: integer, minimum: 0.0 }
        published: { type: integer, minimum: 0.0 }
        processing: { type: integer, minimum: 0.0 }
        dead: { type: integer, minimum: 0.0 }
        failed_attempts: { type: integer, minimum: 0.0 }
        oldest_age_seconds: { type: integer, minimum: 0.0 }
    QueueStatusModel:
      type: object
      required: [realtime_code, bulk_code, balance, threshold]
      properties:
        realtime_code: { anyOf: [{ type: string }, { type: 'null' }] }
        bulk_code: { anyOf: [{ type: string }, { type: 'null' }] }
        balance: { anyOf: [{ type: integer }, { type: 'null' }] }
        threshold: { type: integer }
    QueueResumeModel:
      type: object
      required: [resumed_batches, paused_codes]
      properties:
        resumed_batches: { type: integer }
        paused_codes: { type: array, items: { type: string } }
    ReplayResultModel:
      type: object
      required: [processed_items]
      properties:
        processed_items: { type: integer }
...
    Error:
      type: object
      required: [code, message, detail]
      properties:
        code: { type: string, description: 见 CLAUDE.md 平台错误码表 }
        message: { type: string }
        detail: { type: object, nullable: true }
...
    ExportTask:
      type: object
      required: [id, status, decrypted, row_count, download_url, expires_at, created_at]
      properties:
        id: { type: string, format: uuid, description: 不可枚举公开任务ID；内部整数主键永不对外 }
        status: { type: string, enum: [pending, running, done, failed] }
        decrypted: { type: boolean }
        row_count: { anyOf: [{ type: integer }, { type: 'null' }] }
        download_url: { anyOf: [{ type: string }, { type: 'null' }], description: done且未过期时返回鉴权下载地址 }
        expires_at: { anyOf: [{ type: string, format: date-time }, { type: 'null' }] }
        created_at: { type: string, format: date-time }
...